+++ tests/precision_known_negative_exponent_failure_clean.adb
@@ -1,0 +1,5 @@
+procedure Precision_Known_Negative_Exponent_Failure_Clean is
+   Y : Integer;
+begin
+   Y := 2 ** 3;
+end Precision_Known_Negative_Exponent_Failure_Clean;