+++ tests/precision_known_negative_exponent_failure_dynamic_clean.adb
@@ -1,0 +1,7 @@
+procedure Precision_Known_Negative_Exponent_Failure_Dynamic_Clean
+  (N : Integer)
+is
+   Y : Integer;
+begin
+   Y := 2 ** N;
+end Precision_Known_Negative_Exponent_Failure_Dynamic_Clean;