+++ tests/precision_known_negative_exponent_failure_finding.adb
@@ -1,0 +1,5 @@
+procedure Precision_Known_Negative_Exponent_Failure_Finding is
+   Y : Integer;
+begin
+   Y := 2 ** (-1);
+end Precision_Known_Negative_Exponent_Failure_Finding;