+++ tests/precision_known_negative_exponent_failure_float_clean.adb
@@ -1,0 +1,5 @@
+procedure Precision_Known_Negative_Exponent_Failure_Float_Clean is
+   Y : Float;
+begin
+   Y := 2.0 ** (-1);
+end Precision_Known_Negative_Exponent_Failure_Float_Clean;